pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ascii-forger"
version = "0.2.0"
description = "Convert images into ASCII art — terminal, color, braille Unicode, HTML export, and invert mode."
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Arsh Sharan" }
]
keywords = ["ascii", "art", "terminal", "cli", "image", "image-processing", "ascii-art", "braille", "unicode"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=9.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.0",
    "twine>=4.0",
]

[project.scripts]
ascii-forger = "ascii_forge.main:main"

[project.urls]
Homepage   = "https://github.com/ArshSharan/Ascii-Forge"
Repository = "https://github.com/ArshSharan/Ascii-Forge"
"Bug Tracker" = "https://github.com/ArshSharan/Ascii-Forge/issues"

[tool.setuptools.packages.find]
include = ["ascii_forge*"]
